Update default.logic
Added HasDamageSource for for areas that require killing something (other than darknuts because those only get stunned by bombs)
Area change for HyruleWellLeft as it looked wrong.
Modification to AccessCrenel as you can climb up to crenel fine with a bottle, gust jar and rocscape.
Change To CrenelCaveDownstairs as grip ring is not required for that and the bottle is required for AccessCrenel already.
Change to Greyblade chests as gripring is always required to get to that but the player might not have it.
Change to CrenelBlockChest as current logic had an impossible state with not being able to activate the bridge with grip ring+gustjar.
Renamed TreasureCave to HyliaCapeCave to make its location more apparent, removed a chest that is out of bounds looking at data.
Added FallsLowerCaveRight as it is right next to the other chest.
Modified CloudSouthLeft as it is encased with a cloud wall that the cape wont get you through.
Switched TowerBottomLeft/Right.
Commented out DeepwoodBasementNorth and set the switch chest to only be accessable with rocs cape until the lilypad bug is resolved.
With the removal of DeepwoodBasementNorth I am not sure if another chest needs to be changed to a dungeonItem or not.
Changed CoFAccess to require a damage source and the same logic as to get the CrenelBlockChest.
Changed FortressOutside to FortressOutSideBombWall to make the location more apparent.
Added a commented list of unrandomized or obscure locations that could or could not get get added.