Make formatting consistent. Remove unused usings.
This ensures the code files have a clean, consistent look and feel.

* Tabs vs spaces are resolved to spaces.
* Braces follow the Allman style throughout.
* Any unused using statements were removed, and the remaining ones were sorted, System ones first.

Some more changes are aimed for the future, but I wish to start small.

While an `.editorconfig` file was used to ensure consistency, that file is not ready for distribution. There are some other style issues that may need agreeing upon before I go ahead with any other proposals.