Prep his solution for multiple projects.
This moves all of the source files we created into a more traditional `src/MinishRandomizer` folder. This is done to allow for potential project refactoring or additions. I know there are some that may want a web patch generator or something more flexible: this would eventually allow for UI decoupling and potential unit testing.

The `sln` file was not moved since there are some disagreements from what I have seen on whether said file should be in the project root or in the `src` folder. I will defer on that until there is more of a concensus.