Resolve Dungeon location types
Compass and Map defined in the same way Keys are to enable them to be shuffled to prize locations in their own dungeon if there is no element.

Also trim some of the unnecessary defines for some key locations, the dungeon pots, and a mistake in the underwater HP name.